@@ -10,7 +10,7 @@
     ],
     "license": "GPL-3.0-or-later",
     "require": {
-        "php": "^7.4|^8.0",
+        "php": "^8.0",
         "ext-bcmath": "*",
         "ext-json": "*",
         "ext-ldap": "*",
@@ -19,6 +19,7 @@
         "ext-pdo_mysql": "*",
         "ext-zip": "*",
         "adldap2/adldap2-laravel": "^6.1",
+        "celaraze/laravel-response": "dev-master",
         "celaraze/pour": "dev-master",
         "dcat/easy-excel": "^1.0",
         "dcat/laravel-admin": "2.0.22-beta",
@@ -36,12 +37,12 @@
         "tymon/jwt-auth": "^1.0"
     },
     "require-dev": {
-        "facade/ignition": "^2.5.7",
-        "fzaninotto/faker": "^1.9.1",
-        "laravel/sail": "^1.4",
-        "mockery/mockery": "^1.3.1",
+        "facade/ignition": "^2.5",
+        "fakerphp/faker": "^1.9.1",
+        "laravel/sail": "^1.0.1",
+        "mockery/mockery": "^1.4.2",
         "nunomaduro/collision": "^5.0",
-        "phpunit/phpunit": "^9.3"
+        "phpunit/phpunit": "^9.3.3"
     },
     "config": {
         "optimize-autoloader": true,
