@@ -19,7 +19,7 @@ class DeviceRecordBatchCreateUpdateTrackAction extends BatchAction
      *
      * @return Modal|string
      */
-    public function render()
+    public function render(): string|Modal
     {
         // 实例化表单类并传递自定义参数
         $form = DeviceRecordBatchCreateUpdateTrackForm::make()->payload([
