@@ -215,9 +215,9 @@ class PartRecordController extends AdminController
     /**
      * 返回字段排序.
      *
-     * @return mixed
+     * @return array
      */
-    public function sortColumns()
+    public function sortColumns(): array
     {
         return ColumnSort::where('table_name', (new PartRecord())->getTable())
             ->get(['name', 'order'])
@@ -227,11 +227,11 @@ class PartRecordController extends AdminController
     /**
      * Make a show builder.
      *
-     * @param mixed $id
+     * @param int $id
      *
      * @return Show
      */
-    protected function detail($id): Show
+    protected function detail(int $id): Show
     {
         return Show::make($id, new PartRecord(['category', 'vendor', 'channel', 'device', 'depreciation']), function (Show $show) {
             $sort_columns = $this->sortColumns();
