@@ -160,9 +160,9 @@ class ServiceRecordController extends AdminController
     /**
      * 返回字段排序.
      *
-     * @return mixed
+     * @return array
      */
-    public function sortColumns()
+    public function sortColumns(): array
     {
         return ColumnSort::where('table_name', (new ServiceRecord())->getTable())
             ->get(['name', 'order'])
@@ -172,11 +172,11 @@ class ServiceRecordController extends AdminController
     /**
      * Make a show builder.
      *
-     * @param mixed $id
+     * @param int $id
      *
      * @return Show
      */
-    protected function detail($id): Show
+    protected function detail(int $id): Show
     {
         return Show::make($id, new ServiceRecord(['channel', 'device']), function (Show $show) {
             $sort_columns = $this->sortColumns();
