@@ -61,9 +61,9 @@ class ExpirationService
      * @param $item_type
      * @param $id
      *
-     * @return string
+     * @return int|string
      */
-    public static function itemExpirationLeftDays($item_type, $id)
+    public static function itemExpirationLeftDays($item_type, $id): int|string
     {
         $item = null;
         $item = Support::getItemRecordByClass($item_type, $id);
